+++ lib/queries/one/filter_table.ts
@@ -5,6 +5,7 @@ import {MapTable} from './map_table'
 import {GroupTable} from './group_table'
 import {parsePredicate} from '../../parsing/predicate_parsing'
 import {Value} from '../../column_operations'
+import {GetColumnFromTable} from './get_column_from_table'
 
 export class FilterTable<T> {
     private readonly statement: SelectStatement
@@ -35,7 +36,11 @@ export class FilterTable<T> {
         return new SelectTable(this.constructor, this.statement)
     }
 
-    map<U>(f: (table: T) => U): MapTable<T, U> {
+    get<U extends Value>(f: (table: T) => U): GetColumnFromTable<T, U> {
+        return new GetColumnFromTable(this.statement, f)
+    }
+
+    map<U extends Record<string, Value>>(f: (table: T) => U): MapTable<T, U> {
         return new MapTable(this.statement, f)
     }