+++ lib/queries/one/get_column_from_table.ts
@@ -1,0 +1,13 @@
+import {SelectSqlGenerator} from '../../sql_generation'
+import {SelectStatement} from '../../select_statement'
+import {Value} from '../../column_operations'
+import {parseGet} from '../../generation/get_parsing'
+
+export class GetColumnFromTable<T, U extends Value> extends SelectSqlGenerator {
+    constructor(existingStatement: SelectStatement, get: (table: T) => U) {
+        super({
+            ...existingStatement,
+            selection: [parseGet(get)]
+        })
+    }
+}