+++ lib/queries/one/sort_table.ts
@@ -4,6 +4,7 @@ import {MapTable} from './map_table'
 import {SelectSqlGenerator} from '../../sql_generation'
 import {parseOrder} from '../../parsing/order_parsing'
 import {Value} from '../../column_operations'
+import {GetColumnFromTable} from './get_column_from_table'
 
 export type Direction = 'asc' | 'desc'
 
@@ -33,7 +34,11 @@ export class SortTable<T> extends SelectSqlGenerator {
         return new SelectTable(this.constructor, this.statement)
     }
 
-    map<U>(f: (table: T) => U): MapTable<T, U> {
+    get<U extends Value>(f: (table: T) => U): GetColumnFromTable<T, U> {
+        return new GetColumnFromTable(this.statement, f)
+    }
+
+    map<U extends Record<string, Value>>(f: (table: T) => U): MapTable<T, U> {
         return new MapTable(this.statement, f)
     }
 }