+++ lib/queries/two/aggregate_two_tables.ts
@@ -1,18 +1,7 @@
 import {SelectStatement} from '../../select_statement'
 import {SelectSqlGenerator} from '../../sql_generation'
 import {parseAggregation} from '../../parsing/aggregation_parsing'
-
-export type AggregatableColumn<F> = {
-    avg(): F
-    count(): F
-    max(): F
-    min(): F
-    sum(): F
-}
-
-export type AggregatableTable<T> = {
-    [F in keyof T]: AggregatableColumn<F>
-}
+import {AggregatableTable} from '../one/aggregate_table'
 
 export class AggregateTwoTables<T1, T2, K, A> extends SelectSqlGenerator {
     constructor(