+++ lib/queries/two/filter_two_tables.ts
@@ -4,8 +4,8 @@ import {Value} from '../../column_operations'
 import {SortTwoTables} from './sort_two_tables'
 import {SelectTwoTables} from './select_two_tables'
 import {MapTwoTables} from './map_two_tables'
-import {GroupTable} from '../one/group_table'
 import {GroupTwoTables} from './group_two_tables'
+import {GetColumnFromTwoTables} from './get_column_from_two_tables'
 
 export class FilterTwoTables<T1, T2> {
     private readonly statement: SelectStatement
@@ -37,7 +37,11 @@ export class FilterTwoTables<T1, T2> {
         return new SelectTwoTables(this.firstConstructor, this.secondConstructor, this.statement, first, second)
     }
 
-    map<U>(f: (first: T1, second: T2) => U): MapTwoTables<T1, T2, U> {
+    get<U extends Value>(f: (first: T1, second: T2) => U): GetColumnFromTwoTables<T1, T2, U> {
+        return new GetColumnFromTwoTables(this.statement, f)
+    }
+
+    map<U extends Record<string, Value>>(f: (first: T1, second: T2) => U): MapTwoTables<T1, T2, U> {
         return new MapTwoTables(this.statement, f)
     }