+++ test/parsing/map_parsing.test.ts
@@ -3,30 +3,21 @@ import {parseMap} from '../../lib/parsing/map_parsing'
 import {Employee} from '../test_tables'
 import {createAlias, createGet} from '../../lib/column_operations'
 
-describe('parseMap', function () {
-    it('can accept the selection of a single property', () => {
+describe('parseMap can parse an object', function () {
+    it('with a single key', () => {
         assert.deepEqual(
-            parseMap((e: Employee) => e.id),
-            [ createGet(1, 'id') ]
+            parseMap((e: Employee) => ({employeeId: e.id})),
+            [ createAlias(createGet(1, 'id'), 'employeeId') ]
         )
     })
 
-    describe('can parse an object ', () => {
-        it('with a single key', () => {
-            assert.deepEqual(
-                parseMap((e: Employee) => ({employeeId: e.id})),
-                [ createAlias(createGet(1, 'id'), 'employeeId') ]
-            )
-        })
-
-        it('with two keys', () => {
-            assert.deepEqual(
-                parseMap((e: Employee) => ({first: e.firstName, last: e.lastName})),
-                [
-                    createAlias(createGet(1, 'firstName'), 'first'),
-                    createAlias(createGet(1, 'lastName'), 'last')
-                ]
-            )
-        })
+    it('with two keys', () => {
+        assert.deepEqual(
+            parseMap((e: Employee) => ({first: e.firstName, last: e.lastName})),
+            [
+                createAlias(createGet(1, 'firstName'), 'first'),
+                createAlias(createGet(1, 'lastName'), 'last')
+            ]
+        )
     })
 })